scripts/data-prep/requirements.txt
pymongo==4.7.2
opensearch-py==2.6.0
langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-openai==0.1.8
python-dotenv==1.0.1
sentence-transformers==2.5.1